献立取得APIの一連のエラー対応
